+++ NEWS.md
@@ -1,3 +1,8 @@
+# aramappings 0.1.3
+
+Removed package dependencies involving data sets. The data sets used in tests
+and examples are now included in the package.
+
 # aramappings 0.1.2
 
 Fixed installation instructions in README and vignette